Move type interfaces into a new file
